requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altgraph==0.17.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.1.0
bitarray==2.9.2
blinker==1.8.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
cytoolz==0.12.2
dateparser==1.1.8
django-cors-headers==4.4.0
eth-account==0.10.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.0
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.0
greenlet==3.0.3
gunicorn==22.0.0
hexbytes==0.3.1
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
MarkupSafe==2.1.5
multidict==6.0.4
mysqlclient==2.2.4
numpy==1.26.0
oauthlib==3.2.2
okx==1.0.8
packaging==23.2
pandas==2.1.1
parsimonious==0.9.0
paux==1.0.14
pefile==2023.2.7
pendulum==2.1.2
protobuf==4.25.2
pycparser==2.21
pycryptodome==3.19.0
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
PyJWT==2.8.0
PyMySQL==1.1.1
python-binance==1.0.19
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pytzdata==2020.1
pyunormalize==15.1.0
pywin32-ctypes==0.2.2
redis==5.0.1
referencing==0.32.1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rlp==4.0.0
rpds-py==0.17.1
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.0
toolz==0.12.0
tweepy==4.14.0
twilio==8.13.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
urllib3==2.0.5
web3==6.14.0
websockets==11.0.3
Werkzeug==3.0.3
yarl==1.9.2
